+++ qcore/constants.py
@@ -17,6 +17,7 @@ IM_CALC_COMPONENTS = ["geom", "000", "090", "ver", "ellipsis"]
 
 IM_SIM_CALC_TEMPLATE_NAME = "sim_im_calc.sl.template"
 IM_SIM_SL_SCRIPT_NAME = "sim_im_calc_{}.sl"
+IM_SIM_CALC_INFO_SUFFIX = "_imcalc.info"
 
 MERGE_TS_DEFAULT_NCORES = 4
 
@@ -58,6 +59,11 @@ class ExtendedEnum(Enum):
     @classmethod
     def has_value(cls, value):
         return any(value == item.value for item in cls)
+
+    @classmethod
+    def is_substring(cls, parent_string):
+        """Check if an enum's string value is contained in the given string"""
+        return any(item.value in parent_string for item in cls)
 
 
 class ExtendedStrEnum(ExtendedEnum):
@@ -229,6 +235,7 @@ class MetadataField(ExtendedEnum):
     start_time = "start_time"
     end_time = "end_time"
     submit_time = "submit_time"
+    status = "status"    
 
     im_pSA_count = "pSA_count"
     im_comp = "im_components"